backend/pyproject.toml
[project]
name = "deltaone-backend"
version = "0.1.0"
description = "Backend API for DeltaOne LegalTech Platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "psycopg2-binary>=2.9.9",
    "djangorestframework-simplejwt>=5.3.0",
    "django-otp>=1.5.0",
    "qrcode>=7.4.2",
    "python-decouple>=3.8",
    "django-extensions>=3.2.3",
    "django-storages>=1.14.2",
    "boto3>=1.34.0",
    "celery>=5.3.4",
    "redis>=5.0.1",
    "drf-spectacular>=0.27.0",
    "django-axes>=6.0.0",
    "django-ratelimit>=4.1.0",
    "django-debug-toolbar>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# === CRITICAL FIX: Tell Hatchling where your code is ===
[tool.hatch.build.targets.wheel]
packages = ["backend", "apps"]

[tool.hatch.build.targets.sdist]
include = [
    "backend/**",
    "apps/**",
    "manage.py",
    "README.md",
    "pyproject.toml",
]

# === Development dependencies (not needed for build) ===
[tool.hatch.envs.default]
dependencies = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=5.0.0",
    "factory-boy>=3.3.0",
    "ipython>=8.18.0",
    "django-debug-toolbar>=4.3.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
    "types-boto3>=1.34.0",
    "types-redis>=4.6.0",
    "types-pyyaml>=6.0.12",
    "types-python-dateutil>=2.8.19",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args}"
test-cov = "pytest --cov=backend --cov=apps {args}"
lint = "ruff check ."
format = "ruff format ."
typecheck = "mypy backend apps"

# === Ruff configuration ===
[tool.ruff]
line-length = 88
target-version = "py311"
exclude = ["**/migrations/*.py", "tests/"]

[tool.ruff.lint]
select = ["E", "F", "B", "C4", "I", "UP"]
ignore = ["E501"] # Ignore line too long

[tool.ruff.lint.per-file-ignores]
"backend/settings/*.py" = ["F403", "F405", "E402"]
"**/__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = false
known-first-party = ["backend", "apps"]

# === MyPy configuration ===
[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disable_error_code = ["var-annotated", "attr-defined"]
exclude = ["migrations/", "tests/"]

[tool.django-stubs]
django_settings_module = "backend.backend.settings.development"

# === Pytest configuration ===
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings.development"
python_files = ["test_*.py", "*_tests.py"]
testpaths = ["tests", "apps/*/tests"]
addopts = "--strict-markers --tb=short"
